requirements.txt
Django
psycopg2
djangorestframework
pillow
django-cors-headers